rewrite PhoneRepository using Spring Data

+++ dao/src/main/java/com/getjavajob/training/timashovy/socialnetwork/dao/daoimpl/account/PasswordRepositoryImpl.java
@@ -43,7 +43,9 @@ public class PasswordRepositoryImpl implements PasswordRepository {
 
     @Override
     public void delete(Long id) {
-        passwordRepositorySpringData.deleteById(id);
+        if (passwordRepositorySpringData.existsById(id)) {
+            passwordRepositorySpringData.deleteById(id);
+        }
     }
 
 }
